Update `DeepSORT` Update `DeepSORT` for `numpy==1.24.2`

+++ pyppbox/tk_deepsort/detection.py
@@ -27,10 +27,10 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, cls, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=np.float64)
         self.confidence = float(confidence)
         self.cls = cls
-        self.feature = np.asarray(feature, dtype=np.float32)
+        self.feature = np.asarray(feature, dtype=np.float64)
 
     def to_tlbr(self):
         """Convert bounding box to format `(min x, min y, max x, max y)`, i.e.,
